+++ src/References/Entities/EntityInterface.php
@@ -2,9 +2,7 @@
 
 namespace AvtoDev\StaticReferences\References\Entities;
 
-use Illuminate\Contracts\Support;
-
-interface EntityInterface extends Support\Arrayable, Support\Jsonable
+interface EntityInterface extends \Illuminate\Contracts\Support\Arrayable
 {
     //
 }